Consolidate layout and rename Focused method
- Remove layout field from Delegate, use mainLayout.FocusNext/FocusPrev
- Delete internal/tui/layout.go (redundant LayoutView)
- Add FocusManager to internal/tui and teapot packages
- Rename HasFocus() back to Focused() in View interface
