+++ agents/logs/20260123-050704-filelist-session-subscriptions.md
@@ -1,7 +1,7 @@
 # Task: FileList Session Subscriptions and Focus Navigation
 
 **Started:** 2026-01-23 05:07:04
-**Ended:** 2026-01-23 05:30:00
+**Ended:** 2026-01-23 14:45:00
 **Strategy:** Feature (TDD)
 **Status:** Completed
 **Complexity:** Medium
@@ -12,6 +12,7 @@
 2. Only rerender when these keys change
 3. Rename "selection." keys to "tui." keys
 4. Implement focus navigation for LayoutView with FocusNext/FocusPrev
+5. Consolidate `mainLayout` and `layout` into a single structure
 
 ## Progress
 - [x] Explore codebase structure
@@ -23,6 +24,9 @@
 - [x] Add session subscription to FileListView
 - [x] Update app.go to use FocusNext/FocusPrev
 - [x] Leaf views (FileListView, DiffView, ScrollView, TextAreaView) override AcceptsFocus()/FocusNext()/FocusPrev()
+- [x] Consolidate mainLayout and layout - remove LayoutView, add FocusManager to MainView
+- [x] Separate FocusManager code into dedicated files (internal/tui/focus_manager.go, teapot/focus_manager.go)
+- [x] Remove fileList and diffView fields from Delegate, access via mainLayout
 - [x] Test changes - all tests pass
 
 ## Obstacles
@@ -30,13 +34,51 @@
 
 ## Outcome
 Successfully implemented:
+
+### Phase 1: Session Keys and View Interface
 - Session keys renamed from `selection.*` to `tui.*` (tui.fileIndex, tui.filePath, tui.focusedPane)
 - View interface updated with:
   - `AcceptsFocus() bool` - returns false by default on BaseView, true on leaf views
-  - `FocusNext() bool` - panics when AcceptsFocus() is false, properly cycles focus on LayoutView
-  - `FocusPrev() bool` - panics when AcceptsFocus() is false, properly cycles focus on LayoutView
+  - `FocusNext() bool` - panics when AcceptsFocus() is false, returns false for leaf views
+  - `FocusPrev() bool` - panics when AcceptsFocus() is false, returns false for leaf views
   - `HasFocus() bool` - renamed from Focused()
   - Removed `Focusable()` method
 - FileListView now subscribes to session keys: diff.files, tui.fileIndex, tui.focusedPane
-- LayoutView manages focus traversal with Tab (FocusNext) and Shift+Tab (FocusPrev)
-- All relevant tests pass
+
+### Phase 2: Layout Consolidation
+- Deleted `internal/tui/layout.go` (LayoutView was redundant)
+- Created `internal/tui/focus_manager.go` with new FocusManager for TUI:
+  - Tracks focusedIndex and focusedPane among child views
+  - Provides FocusNext()/FocusPrev() without wrap-around
+  - Used by MainView for focus handling
+- Updated `internal/tui/main_view.go`:
+  - Now uses FocusManager internally
+  - Exposes FocusNext(), FocusPrev(), GetFocusedPane(), SetFocusedPane()
+- Updated `internal/app/app.go`:
+  - Removed `layout` field from Delegate
+  - All focus calls now go through `d.mainLayout` only
+- Separated teapot's FocusManager into `teapot/focus_manager.go`:
+  - Moved from widget.go into its own file
+  - Includes ModalKeyHandler interface
+
+### Phase 3: Remove Duplicate View References
+- Removed `fileList` and `diffView` fields from Delegate struct
+- All accesses now go through `d.mainLayout.FileList()` and `d.mainLayout.DiffView()`
+- This eliminates redundant references while keeping typed access through MainView
+
+### Files Changed
+- `internal/session/session.go` - renamed keys
+- `internal/tui/layout.go` - deleted
+- `internal/tui/focus_manager.go` - new
+- `internal/tui/main_view.go` - added FocusManager
+- `internal/tui/filelist_view.go` - added session subscriptions
+- `internal/tui/diffview.go` - focus method updates
+- `internal/app/app.go` - consolidated layout usage
+- `teapot/widget.go` - removed FocusManager code
+- `teapot/focus_manager.go` - new (extracted from widget.go)
+- `teapot/layout.go` - HasFocus() updates
+- `teapot/layout_test.go` - test updates
+- `teapot/scroll_view.go` - focus method updates
+- `teapot/textarea_view.go` - focus method updates
+
+All relevant tests pass.
